+++ app/Http/Kernel.php
@@ -36,6 +36,7 @@ class Kernel extends HttpKernel
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
             \App\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
+            \Modules\User\Http\Middleware\CheckForzarCambioPassword::class,
         ],
 
         'api' => [
@@ -64,8 +65,9 @@ class Kernel extends HttpKernel
         'throttle' => \Illuminate\Routing\Middleware\ThrottleRequests::class,
         'verified' => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
 
-        'permission' => \App\Http\Middleware\CheckPermission::class,
-        'app.access' => \App\Http\Middleware\CheckAppAccess::class,
+        'permission'           => \App\Http\Middleware\CheckPermission::class,
+        'app.access'           => \App\Http\Middleware\CheckAppAccess::class,
+        'forzar.cambio.pass'   => \Modules\User\Http\Middleware\CheckForzarCambioPassword::class,
     ];
     
 }