+++ bootstrap/app.php
@@ -13,9 +13,12 @@ return Application::configure(basePath: dirname(__DIR__))
     )
     ->withMiddleware(function (Middleware $middleware) {
         $middleware->alias([
-            'permission' => \App\Http\Middleware\CheckPermission::class,
-            'app.access' => \App\Http\Middleware\CheckAppAccess::class,
+            'permission'         => \App\Http\Middleware\CheckPermission::class,
+            'app.access'         => \App\Http\Middleware\CheckAppAccess::class,
+            'forzar.cambio.pass' => \Modules\User\Http\Middleware\CheckForzarCambioPassword::class,
         ]);
+
+        $middleware->appendToGroup('web', \Modules\User\Http\Middleware\CheckForzarCambioPassword::class);
     })
     ->withExceptions(function (Exceptions $exceptions) {
         //