+++ config/versiones.php
@@ -1,9 +1,9 @@
 <?php
 
 return [
-    'IEE'           => '1.13.1',
-    'BhagamApps'    => '1.13.1',
-    'User'          => '2.2.2',
+    'IEE'           => '1.14.0',
+    'BhagamApps'    => '1.14.0',
+    'User'          => '2.3.0',
     'Inventario'    => '2.10.5',
     'Apps'          => '1.5.0',
     'CrudGenerator' => '1.1.0',